apacc_sim_toolkit/apacc-sim/requirements_apaccsim.txt
# requirements_apaccsim.txt
#
# Python package requirements for APACC-Sim toolkit
# Compatible with Python 3.10+ on Windows/Linux
#
# Core packages are required for basic functionality
# Optional packages enhance specific features

# Core Scientific Computing
numpy>=1.21.0,<2.0.0
scipy>=1.7.3,<2.0.0
pandas>=1.4.0,<2.0.0
matplotlib>=3.5.0,<4.0.0
seaborn>=0.11.2,<1.0.0

# Data Formats and I/O
h5py>=3.6.0,<4.0.0
pyyaml>=5.4.1,<7.0.0
openpyxl>=3.0.9,<4.0.0  # Excel support
xlrd>=2.0.1,<3.0.0      # Legacy Excel support

# Visualization
plotly>=5.9.0,<6.0.0    # Interactive plots

# Distributed Computing
ray[default]>=2.0.0,<3.0.0  # Distributed execution
redis>=4.0.0,<5.0.0         # Message passing

# Testing and Quality
pytest>=7.0.0,<8.0.0
pytest-cov>=3.0.0,<5.0.0
pytest-xdist>=2.5.0,<4.0.0  # Parallel testing
pytest-html>=3.1.0,<4.0.0   # HTML reports
pylint>=2.12.0,<3.0.0
mypy>=0.910,<2.0.0

# Performance and Profiling
psutil>=5.8.0,<6.0.0      # System monitoring
py-cpuinfo>=8.0.0,<10.0.0 # CPU information
memory-profiler>=0.58.0,<1.0.0
line-profiler>=3.3.0,<5.0.0

# Simulation Interfaces
# Note: CARLA Python API must be installed separately from CARLA installation
# carla==0.9.14  # Install from CARLA distribution
sumolib>=1.12.0
traci>=1.12.0

# Machine Learning (optional, for DRL controllers)
scikit-learn>=1.0.0,<2.0.0
# torch>=1.10.0,<2.0.0  # Uncomment if using PyTorch
# tensorflow>=2.8.0,<3.0.0  # Uncomment if using TensorFlow

# Data Serialization
msgpack>=1.0.3,<2.0.0
pyarrow>=7.0.0,<15.0.0  # Parquet and Feather support

# Documentation
sphinx>=4.3.0,<7.0.0
sphinx-rtd-theme>=1.0.0,<2.0.0

# Utilities
tqdm>=4.62.0,<5.0.0      # Progress bars
click>=8.0.0,<9.0.0      # CLI helpers
colorama>=0.4.4,<1.0.0   # Cross-platform colored output
python-dateutil>=2.8.2,<3.0.0

# GPU Utilities (optional)
# gputil>=1.4.0  # GPU monitoring, uncomment if NVIDIA GPU available

# Security and Code Quality
bandit[toml]>=1.7.0,<2.0.0  # Security linting

# Development Dependencies (optional, for contributors)
# black>=22.0.0,<24.0.0    # Code formatting
# isort>=5.10.0,<6.0.0     # Import sorting
# flake8>=4.0.0,<7.0.0     # Style checking
# pre-commit>=2.16.0,<4.0.0 # Git hooks

# Platform-specific dependencies
# Windows-specific
pywin32>=300; sys_platform == 'win32'

# Linux-specific
# python-prctl>=1.8.1; sys_platform == 'linux'  # Process control

# Pinned sub-dependencies for reproducibility
certifi>=2021.10.8
charset-normalizer>=2.0.0
idna>=3.3
urllib3>=1.26.8,<2.0.0
requests>=2.27.0,<3.0.0

# Version constraints for compatibility
# Ensure these align with CARLA and SUMO requirements
protobuf>=3.19.0,<4.0.0
grpcio>=1.43.0,<2.0.0